extract normalizers
